FIX: Fix last few steps
